Convert SignInText component to TypeScript
Refactored; TypeScript; functional component.
